resolver: split lib.rs into focused modules
Break aube-resolver/src/lib.rs (6504 lines) into five sibling modules,
dropping lib.rs to 2545 lines:

  tests.rs         2745  full test suite
  error.rs          445  Error enum + diagnostic help formatters
  local_source.rs   357  path rebase, tarball read, git/remote resolvers
  semver_util.rs    245  PickResult, pick_version, range cache
  types.rs          206  MinimumReleaseAge, DependencyPolicy, ResolvedPackage

Public API is preserved via re-exports from lib.rs. Remaining bulk is
the single impl Resolver block; splitting that further is left for a
separate pass.
